+++ app/components/Text/paragraph.template.ts
@@ -1,6 +1,6 @@
 import type { Template } from "tinacms";
+import { accessibility } from "../Base/templates/base.advancedConfigs.accessibility.field";
 import { textEfects } from "../Base/templates/base.advancedConfigs.textEffects.field";
-import { accessibility } from "./templates/text.advancedConfigs.accessibility.field";
 import { fontWeight } from "./templates/text.fontWeight.field";
 import { fontSize } from "./templates/text.size.field";
